test: AbstractTestContainersTest added
