--- main/src/test/kotlin/com/ip13/main/utils/EntitiesProvider.kt
@@ -1,5 +1,0 @@
-package com.ip13.main.utils
-
-object EntitiesProvider {
-
-}
